Refactor YouTube URL extraction method to utils.py file
The YouTube URL extraction method has been moved from youtube.py to utils.py, using the regular expression library that was imported. Also, the subtitle file search in youtube.py has been simplified by removing the unnecessary return statement and directly raising an exception when the file is not found. This clarify the code structure and improve error handling.
